+++ LuiExtendedWelcome.lua
@@ -5,67 +5,51 @@ local windowManager = WINDOW_MANAGER
 
 -- TODO: Make some vars here for message text from changelog.
 -- TODO: Create a basic logo to use when the changelog is displayed - may consider attempting to center it and remove ability to expand box size.
---local memes = zo_iconFormat("LuiExtended/media/changelog/updatenotestemp.dds", 256, 256)
 
 local fillMessages = {
     "|cEEEE00General Components|r",
-    "• Implemented an ingame changelog that will display on the first load with LuiExtended enabled when the version number is different than your current version. This changelog can also be viewed from the addon settings menu.",
-    "• Added the option to switch between using ACCOUNT WIDE or CHARACTER SPECIFIC savedVariables settings. Profiles can be copied between characters or deleted from the addon settings menu.",
-    "• Separated Slash Commands & Info Panel modules into their own individual component menus.",
-    "• Updated and cleaned up various menus to be more intuitive.",
-    "• Replaced GetSynergyInfo() hook with ZO_Synergy:OnSynergyAbilityChanged() hook in order to fix compatibility issue with Innocent Blade of Aoe addon by dorrino. This is a better way to hook, as only the displayed icon is modified and the original synergy info is preserved.",
-    "• Chat Printing settings have been moved to Chat Announcements - and updated to allow the option to print to individual tabs instead of all tabs. This allows you to have loot, experience, etc notifications only display in one chat tab. Also added a toggle to allow system/notification based messages to bypass this and still appear in all tabs. These settings have been reset to default due to this change.",
-    "• Added an option to choose the color for timestamps prepended to chat messages.",
+    "• Added RU Translation thanks to a ton of effort from KiriX!",
     "|",
     "|cEEEE00Buffs & Debuffs|r",
-    "• Added the ability to track prominent player buffs & debuffs by name or AbilityId. Prominent containers support the ability to display the aura name and a timer bar.",
-    "• Added the ability to blacklist buffs & debuffs by name or AbilityId.",
-    "• Updated auras and improved custom icons for various NPC abilities as well as added new icons and updated auras for many NPC's missing them.",
-    "• Added auras & icons for Scalecaller Crown Crate memento effects.",
-    "• New internal table allows simulated aura display on targets based on name. This allows for the display of buffs such as Radiance on Flame Atronachs and CC Immunity on boss rank enemies.",
-    "• Added a Debug option to display the AbilityId of effects on their buff/debuff icon.",
-    "• Added Night Mother's Gaze & Line-Breaker into debuffs that always display on a target regardless of source.",
-    "• Fixed an unintentional change which modified the Line-Breaker debuff from the Alkosh set to be named \"Roar of Alkosh.\" This should prevent Combat Metrics from combining the two debuffs when reporting uptime.",
-    "• Fixed an issue where Off Balance Immunity wasn't displaying on targets and adjusted it to display as a buff rather then debuff. This effect can still be tracked by prominent buffs & debuffs due to its importance.",
+    "• Did a pass on ALL player skill line abilities - updated auras for all abilities with name corrections, icon corrections, and hid unneccessary dummy auras.",
+    "• Added custom icons for the attacks used by Nightblade, Sorcerer, and Warden pet abilities.",
+    "• Added custom icons for a great deal of NPCs that were missing them including all Dwemer NPCs, Giants, Trolls, Daedric Hungers, Iron Atronachs, etc...",
+    "• Updated localization for major/minor debuffs so that all languages will show major/minor debuffs on the target regardless of source.",
+    "• Reset the options for \"Display Extra Buffs & Consolidate Buffs\" to account for some changes made with this update as well as added an additional option to consolidate single major/minor auras.",
+    "• Grim Focus updated to track its stack count on the duration buff applied to the player.",
+    "• Fixed an issue where Ground Mine stacks weren't decremented when a mine triggered if an aura was set to prominent.",
+    "• Removed a significant amount of older code related to ground effects that should result in minor performance enhancement.",
+    "• Fixed an issue where the Pelinal's Ferocity experience buff was not correctly labeled as an experience buffs for the purpose of hiding it from display.",
+    "• Updated auras for various sets, and added updated icons/auras for sets from Dragonstar/Maelstrom Arena, as well as all Summerset sets.",
+    "• Fixed an issue where prominent auras weren't removed from tracking when the player died.",
+    "• Added the option to display a timer buff when transformed into a Werewolf. This timer will pause at a Werewolf Shrine or when using the Devour passive.",
+    "• Fixed an issue where dummy auras added by target name (example: CC Immunity buff added to some boss NPCs) would display on dead targets.",
+    "• Fixed an issue where Artifical Effect id's would not refresh when updating settings in the Buffs & Debuffs menu (dummy auras added by ZOS - such as Battle Spirit).",
     "|",
     "|cEEEE00Chat Announcements|r",
-    "• Updated the default variable settings for Alliance Point, Tel Var Stone, and Experience Point notifications so throttling is enabled by default.",
+    "• Fixed an issue that was causing Disguise Notifications not to display even when enabled.",
+    "• Updated Achievement Tracking options with all current categories.",
+    "• Added Psijic Order guild reputation tracking.",
+    "• Fixed an issue where the menu option to modify the context currency/loot message for \"Spend\" was missing.",
+    "• Added a new context message option \"Pay\ for gold given to NPC's in conversation options.",
+    "• Jewelry Crafting should now be supported by Loot crafting messages.",
     "|",
     "|cEEEE00Combat Info|r",
-    "• Added an option to toggle on/off the display of bar highlight for secondary effects like the magicka restore from Honor the Dead or Major Savagery from Biting Jabs.",
-    "• Added option to change the formatting of the Ultimate Percentage Label & set the default font/position to match that of bar highlight.",
-    "• Added an option to play a sound on ability procs.",
-    "• Anchored bar labels to Action Button instead of flipcards, this stops the label from wiggling erratically when swapping bars.",
-    "• Fixed an issue where when bar ability highlight for the Ultimate slot was enabled along with Ultimate % label the two labels would overlap on application for a split second.",
-    "• Fixed an issue where the initial bar highlight label created for effects would display the default base ability duration without compensating for the extra duration added to many abilites added by EVENT_EFFECT_CHANGED.",
+    "• Did a pass on ALL player skill line abilities - Bar Ability Highlight tracking is now supported for any ability the player can slot.",
+    "• Bar Ability Highlight tracking now supports multi-target tracking, when changing targets the highlighted abilities will update based on what debuffs are currently applied to that target.",
+    "• Fixed an issue where the Crystal Fragments proc sound wouldn't play due to a bad variable name.",
     "|",
     "|cEEEE00Combat Text|r",
-    "• Added an option to set Transparency for Combat Text.",
-    "• Added an option to abbreviate numbers like 12345 to 12.3k.",
-    "• Updated Incoming Ability Alerts to hide suggested mitigation options by default. Now messages will simply display incoming ability name/icon unless mitigation method notification is toggled on.",
+    "• Fixed an issue where errors would be displayed when the Shorten Numbers option was enabled.",
     "|",
     "|cEEEE00Info Panel|r",
-    "• Info Panel module has been moved to its own \"LuiExtended - Info Panel\" settings menu.",
-    "• Fixed an issue where the Info Panel displayed over other UI menu's by adding it into the Scene Manager.",
-    "• Fixed an issue where the \"Feed Now\" timer was clipped.",
-    "• Fixed an issue where the Armor Durability indicator was clipped.",
-    "• Removed outdated Imperial City Trophy Display & Scaling Options.",
+    "• Reimplemented the menu option to change the scaling on the Info Panel for higher resolution displays.",
+    "• Mount feed cooldown notification updated from \"Feed Now\" to \"Train Now.\"",
     "|",
     "|cEEEE00Slash Commands|r",
-    "• Slash Commands module has been moved to its own \"LuiExtended - Slash Commands\" settings menu.",
-    "• Slash commands can now be enabled or disabled individually.",
-    "• Added /banker, /merchant, /fence commands to summon assistants.",
-    "• Added /ready command to initiate ready checks.",
-    "• Added keybinding options for /banker, /merchant, /fence, /ready, /home & /regroup.",
-    "|",
-    "|cEEEE00Unit Frames|r",
-    "• Added an option to display the AVA Icon & Rank Number independently from the Title or AVA Rank Name on Target Frames.",
-    "• Added an option to set the priority for AVA Rank vs Title Display on Target Frames.",
-    "• Added individual options to choose the Low Resource threshold for HP/Magicka/Stamina labels.",
-    "• Added an option to select the transparency of the shield bar when using the default Overlay mode.",
-    "• Fixed an issue where the AvA rank icon of players of the same faction was colored white. The AvA icon for players of the same faction will now display the proper faction color.",
-    "• Fixed an issue on Target Frames where rank 0 \"Citizen\" players would display with a rank number of \"ava.\"",
-    "• Fixed an issue where changing the Player Frames layout would also reset the position of Group, Raid, Boss, and AvA frames.",
+    "• Added the \"/outfit\" \"#\" command. Allows you to swap to an outfit. Also adding keybinding options.",
+    --"|",
+    --"|cEEEE00Unit Frames|r",
 }
 
 function LUIE_WelcomeScreen(menu)
@@ -101,15 +85,6 @@ function LUIE_WelcomeScreen(menu)
             luiChangeLog:AddText(strformat("|c00C000If you have any feedback, bug reports, or other questions about <<1>> please visit ESOUI (http://www.esoui.com/downloads/fileinfo.php?id=818), GitHub (https://github.com/ArtOfShred/LuiExtended/) or contact ArtOfShred at artofshred@artofshred.net.|r", LUIE.name))
             luiChangeLog:AddText("|")
 
-            --[[
-            luiChangeLog:SetSlider(23)
-            function luiChangeLog:SetSlider(position)
-                self:GetNamedChild("Buffer"):SetScrollPosition(position)
-                local sliderValue = self:GetNamedChild("Buffer"):GetNumHistoryLines()
-                self:GetNamedChild("Slider"):SetValue(sliderValue - position)
-            end
-            ]]--
-
             -- Create extra backdrop to make the window more visible
             luiChangeLog.extrabackdrop = LUIE.UI.ChatBackdrop( luiChangeLog, nil, nil, nil, 32, false)
             luiChangeLog.extrabackdrop:SetAnchor(TOPLEFT, luiChangeLog, TOPLEFT, -8, -6)
@@ -133,8 +108,8 @@ function LUIE_WelcomeScreen(menu)
 
         local buffer = luiChangeLog:GetNamedChild("Buffer")
         local slider = luiChangeLog:GetNamedChild("Slider")
-        buffer:SetScrollPosition(48)
-        slider:SetValue(buffer:GetNumHistoryLines() - 48)
+        buffer:SetScrollPosition(30)
+        slider:SetValue(buffer:GetNumHistoryLines() - 30)
 
     end
     -- Set version to current version.